OpenGL sample ported to 64-bits

+++ OpenGL/OpenGLES20Example/GLViewController.cs
@@ -1,6 +1,6 @@
 using System;
 using OpenTK;
-using MonoTouch.UIKit;
+using UIKit;
 using OpenTK.Graphics.ES20;
 
 namespace OpenGLES20Example
@@ -207,8 +207,8 @@ namespace OpenGLES20Example
 			modelViewMatrix = GLCommon.Matrix3DMultiply (translationMatrix, rotationMatrix);
 
 			GLCommon.Matrix3DSetPerspectiveProjectionWithFieldOfView (ref projectionMatrix, 45.0f, 0.1f, 100.0f,
-			                                                          View.Frame.Size.Width /
-			                                                          View.Frame.Size.Height);
+				(float)View.Frame.Size.Width /
+				(float)View.Frame.Size.Height);
 
 			matrix = GLCommon.Matrix3DMultiply (projectionMatrix, modelViewMatrix);
 			GL.UniformMatrix4 (matrixUniform, 1, false, matrix);
